Create a controller for the initial and runtime frequency tuning
The objective of this variable is to hold the frequency range
over which we will try to tune.

It is used with a function probe that calls its get_next_freq function.

Since this object also knows when the frequency search has finished
it may also be used to hold the collected timing information.

In this scenario the function probe can be used to both set the last
synchronization information and get the next center frequency

Also get and sort the timing_metric from frame-sync

CHANGELOG:
  - Create a function in the frame_sync_fast to return the average timing metric
  - Store the timing information into frequency tuner class
  - Save the store timing/frequency metrics into a csv when sweep finishes
  - Make the frequency tuner able to stop the flow-graph
  - Evaluate the timing metric even wihout a detected peak
  - Pre-calculate the frequency range to be tested
  - Add test script for initial tuning in real-time
  - Add command line arguments to configure the freqency range and sweep time