Force dependency of Tuning Control block w.r.t. the RTL module
The function that sets the RF center frequency within the Tuning
Control class now receives the current RF center frequency as
parameter. The Function Probe block, in turn, passes to this
argument the result obtained after the call to the RF center
frequency getter from the RTL SDR module. This effectively forces
the RTL SDR to be instantiated before that in the generated
python code and, ultimately, prevents the bug that was ocurring
due to the RTL SDR module being instantiated past the Function
Probe.