Add MER measurement at decoder input
Add MER measurement taken based on the symbols at the input to the
constellation de-mapper. This MER has a more direct relation with the
error performance that is achieved.

Note that previously only the MER at the input to the frame
synchronizer was displayed in the SNR tab of the GUI. The motivation
was that this MER is always available, even when frame sync is not yet
acquired, which is good for antenna positioning. Now, in order to be
able to debug error performance more consistenyly and also allow
positioning, both MER measurements are now displayed.