Change PMF threshold that validates timing metric
The frame timing metric is normalized by the signal energy. Thus, in
the absence of signal, it is prone to being very high (as a result of
normalizing by a close-to-zero value). The magnitude of the PMF output
is used to validate the timing metric in this case, since the PMF
output is always normalized by a pre-defined (theoretical) energy, not
the energy of the actual signal.

The previous PMF magnitude threshold was still not enough under low
SNRs, as noise alone could be enough for the threshold to be
exceeded. Increase the threshold now. Note the value is based on
experiments, so it is likely not be the best choice yet.

This change is motivated by the goal of printing the timing metric
more accurately in the console during initialization. It is especially
suitable when still positioning the satellite dish.