Improve peak detection used for carrier recovery
The carrier frequency recovery scheme relies on a technique that finds
the peak of an FFT. The previous implementation searched for the peak
by using an argmax GR block. However, the limitation was that, under
very low SNRs, the entire FFT vector could converge to a similar
value, with no clear peak. In that case, argmax would still return a
peak, but this peak would be wandering around and would be merely due
to noise.

The new implementation now considers both the maximum and the second
maximum value in the FFT, in order to infer whether there is a
clear (prominent) peak in the FFT vector. To do that, a customised
peak finder was introduced. There is a threshold for the difference
between the maximum and second maximum, which is considered internally
before the peak finder outputs the index of the peak.