Prevent noisy CFO estimations in runtime
The runtime CFO controller checks whether a CFO estimation is close
enough to the value that the estimation has converged to during the
system's operation. This was previously taken into consideration only
for the physical adjustments (to the RF chip) carrier frequency. That
is, before adjusting the hardware tuner, first the block checks if
this adjustments comes from a reliable estimation (one that it has
converged too). Now, this is similarly checked for every digital CFO
correction value too, with the goal of preventing the CFO estimation
from wandering when the SNR is too low. This means that, before
outputing a CFO estimation, first the system ensures it is reliable.