Add Makefile
Support the build/installation of dependencies (gr-mods and gr-framers
blocks) and the generation of the top-level Python applications
corresponding to the .grc flowgraphs. Also support the build of
hierarchical blocks, in case needed (for development). Use this
Makefile as a substitute to the previously used .sh installation
scripts.

Since now the Makefile handles the generation of python applications
via grcc, remove pre-generated python files from revision. One
motivation is that grcc generates python files differently depending
on the user's GNU Radio version. Therefore, using pre-generated python
files should be avoided.

Lastly, since the Makefile also handles the build of hierarchical
blocks, reintroduce the source for these, namely their .grc files.
Note, however, that in contrast to the Rx flowgraphs, the generated
python files of the hier blocks are already included in the repo, so
by default their corresponding grcs will not trigger a new build (they
are already built).  The rationale is that there are some
customizations to their python files, so generation via grcc alone is
not sufficient to obtain what the Rx applications expect.

Remove rx run