Integrate new Tuning Control in Rx flowgraphs
Remove the function probe module (not necessary anymore) and
update/add parameters used by the new implementation of the Tuning
Control module. For example, parameters that define the number of scan
iterations and the duration of each iteration. Default to 20 frequency
scan iterations.

Also set separate work intervals for CFO polling and frequency
scanning. The former is supposed to be quick, to maintain the runtime
tracking of the CFO, while the latter needs to be long for the frame
timing metric to be averaged well.

To support the interpretation of the frequency scanning, display the
RF center frequency that is currently adopted in the SDR board in the
GUI, as a GUI label.

Lastly, delete the flowgraph dedicated for initial tunning, as the
Tuning Control module can now operate in scan mode from the main
flowgraph itself.

Additional changes:

- Update the CFO recovery object used by the logger. Use the top-level
  frequency recovery object within the instances of the Tuning Control
  module, instead of the submodule (the runtime CFO controller).
- Review parameter values and labels in all flowgraphs. Assign short
  ids to all parameters, add descriptions and remove unused
  parameters.
- Set the initial enabled/disabled state of the Rx logger.

- Include optional DC offset filtering in the Rx GUI receiver.