Add scan mode and refactor the frequency tuning
Refactor the "Tuning Control" (TC) module such that it supports the
new "scan mode" feature and becomes the central coordinator of:

1) The scanning/sweeping of RF frequencies prior to the
   initialization, when requested.
2) The frequency corrections made in the SDR hardware whenever the the
   CFO correction range is about to be exceeded.
3) The digital corrections of frequency made via the numerically
   controlled oscillator (NCO).

Task #1

Task #1 is a newly supported mode, based on the (yet unused) previous
implementation of the TC. When enabled, the module sweeps the desired
range of frequencies and, in the end, automatically switch to its
normal operation mode, i.e. with scan mode disabled and role #2 (from
above) enabled. This way, there is no need for a separate application
dedicated to frequency scanning. Also, the Rx application can continue
running without the need for being re-launched.

Importantly, in this new implementation, after the scan, the SDR is
configured with the inferred best RF frequency, where the signal of
interest was found.

During the scan, each frequency is tested for a predefined interval
and the criteria for judging whether it is a suitable frequency is
based on the frame timing recovery metric (whether frame timing can be
acquired). To ensure this works well, the TC updates the "reaction
speed" of the feed-foward CFO recovery module during the scan
phase. That is, it ensures the CFO recovery responds quickly enough,
such that a digital CFO correction is applied in less than the
interval of a scan iteration. As a result, the duration of the scan
iteration becomes sufficient for frame recovery to occur.

To support a GUI label displaying the current RF frequency (useful
during scan mode), add entry in the TC module for the GUI label
object. This ensures dependency of the TC block on the GUI label at
the top-level python. Ultimately, it avoids any conflict with
non-existing label in case the TC calls a frequency update before the
top-level flowgraph instantiates the label object.

Lastly, to ensure that the scan logs are clean and easy to visualize
in the console, disable the Rx logger temporarily from the TC
itself. Re-enable the logger as soon as the scan is complete.

Task #2

The TC module already participated in the process of frequency
corrections in HW (and NCO updates). However, this was previously
dependent on a function probe block, which would launch the thread for
polling the target HW frequency according to the runtime CFO
controller (the RCC) module. Now, this dependency is removed. Polling
threads are launched internally, not only for this control of the
target RF frequency (based on estimations from the RCC), but also for
scan mode and for NCO updates.

To comply with the architecture where the TC is the central point for
frequency corrections, expose new functions in the RCC that allow
explicitly telling the RCC when the HW frequency update is
accomplished. In the RCC, reset the digital CFO correction whenever
this function is called by the TC. This avoids using the digital CFO
correction once the hardware is corrected to clear this error.

Move also the analog HW frequency correction logs from the RCC module
to the TC module. That is, whenever a change to the HW frequency is to
be made, instead of printing when detected (at the RCC), print it when
it is actually executed (at the TC). An important benefit of that is
that, when the tuning control module is performing a frequency scan,
any detection in favor of HW correction at the RCC is ignored (TC does
not process it). In this case, the correction would be printed to the
console (if left in the RCC), but in reality would not be made.

Lastly, update the runtime CFO controller block to use the RF center
frequency as an int variable. This is because the single-precision
float can not resolve numbers such as 1200000020 (1.2 GHz + 20 Hz),
which are practical for the RF center frequency. Besides, the
fractional part of the center frequency is mostly irrelevant.

Task #3

Just like for task #2, remove dependency with respect to a "function
probe" module for NCO updates too. To do so, launch a thread
internally dedicated to NCO updates. This thread polls the target
digital frequency to be used in the NCO and sets the NCO frequency
directly (rather than via a top-level parameter). To support this
behavior, update the NCO module such that it exposes get/set methods
for its target synthesized digital frequency.

The NCO module is now dependent on the target frequency correction and
the sample rate, rather than a phase increment. With these two values,
the phase increment can be computed internally (in the NCO
module). The TC block sets the target frequency for the NCO.

Additional changes:
- Remove dumping of TC results to file.
- Remove verbosity from pipe sink/source, to clear the console from
  unnecessary logs.