Re-introduce the BE FLL and review parameters
Substitute the NCO by the band-edge (BE) frequency-locked
loop (FLL). The NCO is part of the feed-forward (FFW) coarse carrier
frequency offset correction method. However, there is currently a
performance issue in the NCO architecture in that it is challenging to
ensure NCO updates only during the frame preamble. This leads to
frequency corrections in the middle of the frame and, accordingly,
sporadic bit errors, which ultimately result in HDCL frame drops.

To revert into a stable version in terms of BER performance,
re-introduce the BE FLL. Accordingly, review the set of parameters to
optimize the performance and remove parameters corresponding to the
FFW method.

More specifically, review:

- Costas loop bandwidth: the Costas loop serves as a fine frequency
  correction stage, for either frequency correction scheme. It is
  particularly useful to enable the frame recovery. To support
  low-stability LNBs, it should be more responsive, which is achieved
  with wider band (despite noisier behavior).

- CFO poll rate: this is only used for the runtime updates to the
  hardware center frequency in response to CFO exceeding the
  correction range. There is no need to poll fast, as the oscillator
  changes relatively slowly. Hence, the rate is now reduced to once
  per second.

Remove:

- The CFO averaging alpha parameter.
- Frequency recovery FFT size parameter.

Other changes:

- Launch NCO thread only if NCO object is passed to the Tuning Control
  block. This allows the infrastructure for the NCO to remain in
  place, but also not necessarily be used.

- Adapt the Tuning Control module to speak to the FLL and use it for
  the runtime frequency control. For example, to get its frequency
  offset estimation and reset it during a HW freq. update. Add the
  sample rate and oversampling ratio as parameters to the Tuning
  Control block, so that the CFO threshold can be computed according
  to the FLL pull-in range.

- Consider the FLL CFO estimation during the exit routine of the scan
  mode.

- Adapt the RX logger to print CFO estimations from the FLL.

- Adapt the scan mode range and step in compliance to the FLL pull-in
  range.