Reduce the oversampling ratio in all applications
Use an oversampling ratio of 6, instead of 8, with the goal of
reducing the CPU utilization of all receiver applications that include
the low PHY. Also remove the restrictions on the symbol rate that were
previously imposed in the flowgraph to comply with USRP sample rate
restrictions. These are not needed for the RTL-SDR receiver. Lastly,
make the oversampling ratio a program parameter that can be changed
via argument.