Add make install rules and refactor make
Introduce Makefile rules to install the satellite receivers as well as
the gr-mods and gr-framers dependencies. Define separate installation
rules for each of them, due to the fact that gr-mods and gr-framers
need to be **installed** before the satellite **build**.

The gr-mods and gr-framers modules were previously **installed**
during the satellite **build** itself. The main problem with this
approach was that "sudo make install" was called from inside the
Makefile itself, preventing a user to install these modules without
using "sudo". This was a problem e.g. for developing the RPM package,
where the "make install" rule is not executed with root
permissions. To overcome this, now gr-mods and gr-framers dependencies
are installed by calling standalone Makefile install
rules (install-framers and install-mods). This also allows them to be
installed on a target DESTDIR, if specified.

With regard to the satellite receiver applications, rename them such
that "bs_"" is used in front to avoid confusion with other
applications in the OS. **Build** them on a "build" folder (rather
than the project's root folder) and set rule in Makefile
to **install** these generated python applications within "/usr/lib"
or a target destdir. Furthermore, compile the python applications and
create short wrapper scripts to call the byte-compiled
versions. Install the wrapper scripts within "/usr/bin". The wrappers
are named with hyphen rather than underscores, so that ultimately the
installed "applications" (the wrapper scripts) are named e.g. "bs-rx"
or "bs-rx-gui".

This patch also introduces other minor changes into the Makefile:
- Set the shell to bash. This is needed for the adopted string
  manipulation.
- Remove the "remove" rule, leaving only "clean" and "uninstall"
  rules. Both of these now have three different versions,
  e.g. uninstall, uninstall-framers and uninstall-mods.
- Ensure to clear the build record on gr-mods and gr-framers when
  these are uninstalled.