Prevent module issue during daemon thread teardown
When the parent process of a daemon thread stops and the thread is
sleeping, it is apparently still process that it executes shortly
after waking up. However, the thread won't be able to resolve modules,
in which case the execution of "time" module fails in the files being
changed. This is discussed in:

https://stackoverflow.com/a/21844166