Set gr-framers as submodule
Instead of cloning the gr-framers repository from the Makefile, set it
as a submodule of this project and assume the clone already exists
when running the Makefile.

This change also impacts the Makefile build strategy. The adopted
approach generates a single "build_record" file output that is used to
mark the build timestamp. Since gr-framers is now a submodule,
generating such a file inside the repository would lead to a "dirty"
submodule, that is, with untracked changes. Hence, the gr-framers
build record path was moved to the root "build" folder.

framers record