Preserve the build folder on make clean
The result of removing the build folder can be ignored in the
Makefile (with dash prefix in the line) or the command can be removed
entirely. This is necessary in order to ensure the "make clean"
rule does not fail when there is nothing built yet. We choose the
latter approach of removing the line.