Fix path to installed Rx apps in shell wrappers
The .sh wrappers were pointing to a path based on "DESTDIR". While this works
for a local install, where DESTDIR is "/usr/local", it fails e.g. when building
a .deb package, when DESTDIR is a temporary path. The $libdir variable should
already contain the "prefix" and be enough for both cases. Hence, this patch
removes $DESTDIR.