Add examples for sending and receiving API data
Add a sender and a receiver script to illustrate the transmission of data via
the Ionosphere API. The sender script encapsulates a file specified via command
line argument into a user-specific message format with GnuPG encryption, asks
for the user bid and sends the transmission request to the API. The receiver
script decrypts and parses the received data in order to retrieve the file.

Describe the API transmission use cases in a README. Explain the use case in
which the sender/reader scripts are used to transmit files directly as they are,
without any application-specific data structure, in addition to the
user-specific protocol encapsulation use case.

Add also an example demo receiver that is able to fetch data directly from the
Ionosphere API via the internet and that imitates the pipe output of the actual
blocksat receiver.

Add the pipe module that is available in gr-blocksat directly in the examples
directory. This avoids the requirement of building gr-blocksat prior to testing
the API transmission.