Remove Bitcoin FIBRE submodule and update README
The FIBRE submodule was previously used solely to indicate the version that is
used by Blockstream Satellite. Since the FIBRE submodule was not a dependency
for building the satellite receiver application, this patch removes it and
instead clarifies the version that is supposed to be used in the README only.