Reorganize GRC flowgraphs using 3.7.11 format
At the time of this patch, most users rely on GNU Radio version 3.7.11, as it is
the one distributed for Fedora 27/28, Ubuntu Bionic and CentOS 7. Hence, to
facilitate, reorganize all flowgraphs in compliance to the 3.7.11 format.

Also, improve the organization of parameters in the flowgraph. Make sure they
all have the parameters in the same order.