Change first GUI tab to an overview of metrics
Modify the first tab of the GUI such that, instead of showing solely the two MER
estimations, it gives an overview of metrics that can be of interest. In
particular, this new Overview page shows the spectrum, the spectrogram, the
constellation, the signal RMS value (entitled "level"), in addition to the
pre-existing MER estimations.

These changes rely also on an update to the Rx Logger block from the gr-blocksat
library. Hence, this patch also updates the pointer to the gr-blocksat
submodule.