Update and add command-line MER average parameter
Update the instances of the MER measurement block such that they use the new
alpha parameter to control the IIR averaging, as a replacement for the previous
N that controlled the length of the FIR averaging window.

Additionally, include a parameter block to allow configuration of the alpha
value via command-line argument.