Change default DA phase recovery loop bandwidth
Update the default loop bandwidth used by the data-aided carrier phase
recovery. Now that the Costas Loop block is not used anymore, the phase recovery
has to become more responsive in order to cover the setups where phase noise is
relatively high. The value of 0.01 seems to be an intermediate value within the
range of loop bandwidths that are appropriate for varying levels of phase noise
from different LNBs.