Default to using the CFO sleep mechanism
To save some CPU usage, default to using the CFO sleep mechanism, in which the
CFO recovery block estimates and updates the frequency offset correction
intermittently. The sleep period controls the number of FFT blocks that are
waited in order to output a single frequency offset estimation. Instead of
outputting a new offset on every FFT block, as before, output once after every
20 FFT blocks.