Replace protocol version arg by software version
Remove the command-line argument that previously defined the protocol version,
which was used during the transition from version 1 (no encapsulation) to
2 (encapsulation on Blocksat Packets). The motivation is that only version 2 is
supported now, so there is no need to control the protocol version anymore.

Replace the `--version` command-line argument with a command that instead prints
the software version. Use the Makefile to replace a line on the auto
GR-generated Python file and include the software version according to latest
git tag.