Remove "prompt for exit" on console-only apps
In order to allow running blocksat-rx as a service or in the background, change
the run mode to "run to completion", instead of using "prompt for exit". The
problem comes from the fact that the latter mode will execute a "raw_input" for
user to press enter, which then prevents proper background execution. With "run
to completion", there isn't a prompt for user input and the service can still be
killed normally.