Move order of zap command and exit gracefully
Execute zap as the last step of the launch command. The main motivation is to be
able to remove piping of stdout and stderr of the subprocess Popen command in
order to handle prints. The dvbv5-zap application divides prints to both stdout
and stderr, in addition to frequently clearing the output buffer. As a result,
it becomes complicated to handle stdout and stderr buffers from subprocess.

Add also handling of SIGINT signal. Terminate the zap application when a SIGINT
is sent.