Revert read_cfg_file arguments
Passing the entire argparse Namespace was a workaround, but not the best
solution. It is better to provide to read_cfg_file solely the arguments of
interest. In case a configuration file does not exist, create an argparse
namespace containing the configuration directory and file, and then call the
configure function.