Support monitoring of the S400 through SNMP
Add command to fetch the receiver statistics from the S400 via SNMP. Print this
information to the console using the Monitor class from monitor.py.

Use the SNMP implementation from the pysnmp package and add it to the package
setup recipe. Add also all the required MIB files so that the compiled version
of the S400's MIB can be generated locally with no need to fetch MIBs from a
remote server.

On initialization, fetch the configurations from the S400 and display them on
the console.