Reapply static IP config even if IP is right
Now that a configuration file is created in order to set the static IP address,
we must make sure that this file is generated. There is a chance the IP will be
right already after creating a dvbnet interface that was just removed. In this
case, the previous code would not generate the interface configuration
file. This patch fixes this issue.