Install leandvb and ldpc_tool from binary package
Completely remove the compilation of SDR software from source. Install leandvb
and ldpc_tool from the leandvb binary package that is now available on your
repository.

Update the dependency installation tool such that it does not create the
directories for building software from source. Also, remove the installation of
build dependencies required for leandvb and ldpc_tool, as they are not necessary
anymore now that the binaries are shipped directly.