Move checking of deps to dependencies module
The advantage of concentrating everything there is that we can change the
behavior easily in the future. We could for example ask to run the dependency
installation routine automatically.

In this patch, in addition to moving the checking of dependencies for the sdr
module, add also some new checking regarding USB and standalone Rx dependencies.