Support concatenation of bitcoin.conf configs
Add option to concatenate new options into the auto-generate bitcoin.conf
file. This is meant to support the dual-satellite setup, for which the user
needs to add two different sets of configurations to the same bitcoin.conf file.

To avoid duplicate configurations, handle the parsing of bitcoin.conf options
through a dedicated class. The class can load configs from text into a
dictionary and vice-versa. When concatenating new options, it checks if the
options already exist in the current configurations.

Add also some unit tests for the new configuration class of the bitcoin.py
module.