Fix mutual exclusion between RTL index and gain
The mutual exclusion should be between argument -f and all other RTL parameters
such as -g, `--sps`, and --rtl-idx, because when the IQ file is read from a
file, all RTL options are bypassed. Since this is not possible via argparse,
remove the mutual exclusion group completely and define the arguments normally.