Clean old/timed-out messages on API listener
Sometimes the listener does not receive all fragments that are necessary to
decode an API message. In this case, the fragments remain within the packet
handler (i.e., the BlocksatPktHandler object). As a result, the handler's
fragment map can grow over time. To solve this, add a timeout and clean up
timed-out messages. Set the timeout to 2h by default. This value is sufficiently
larger than the maximum transmission delay, which is currently of 17.5 minutes
for the maximum API message size of 1 MB.