Check if the CLI has updates from every command
The previous implementation searched for CLI updates only from commands within
the dependencies.py module. Given that this module is only executed during the
initial setup, the result could be that the user would not search for CLI
updates frequently enough.

This patch moves the update verification logic to a separate module, named
update.py. It now calls the update verification directly from the main.py
module, such that it runs before any CLI command.

Because the verification of available pip updates can be slow, the new
implementation checks pip on a thread, in a non-blocking way. It also maintains
an update cache file, which can be checked quickly before checking pip.