Wait for order transmission on api send command
Instead of returning soon after the "api send" command, keep polling the state
of the order until it is canceled or successfully transmitted.

Add method on ApiOrder object to allow waiting for a target state (or
states). Reorganize also the ApiOrder implementation such that the uuid and
authentication token are saved into the object after transmission. Later on,
this information is used when polling the server for the order's state.