Move common logging args to the monitoring module
Instead of defining the same logging options on the ubs, standalone, and sdr
modules, define them once in the monitoring module. Add a function on the
monitoring module to add these args to a parent parser.

For consistency, let option `--log-interval` exist for all receivers. Use it
internally within the monitoring module to restrict the logging rate. With that,
the log interval logic used on sdr.py can be removed.