Support FEC encoding of large messages
Support the encoding of messages exceeding the 256-chunk capacity of a single
FEC object. Use multiple objects to encode the message and convey enough
metadata on the resulting FEC packets so that the receiver knows how many
objects and chunks to process.

Move the implementation to a dedicated module (fec.py). Add also an independent
set of unit tests just for the FEC module.