Fix bufsize used for socket recvfrom
Bufsize should be a small power-of-two size. The large number adopted on net.py
was a legacy from the original goal of using fragmentation at IP level. Since
fragmentation at IP level was not used in the end, there is no need to keep the
large number. A small bufsize works better.