Detect FEC messages automatically on api listener
The FEC messages hold a unique structure. For every incoming packet, the
listener app verifies whether the structure is decodable already. When it is,
there is a very high chance that the message is really FEC-decoded. If all
fragments of a message are received and the message remains not ready to be
FEC-decoded, it can be inferred that the message was not FEC-encoded at
all. Hence, there is no need for a user-specified command-line
argument ("--fec"). The chances of having a non-FEC message whose contents
coincide with the expected FEC structure are negligible.