Shuffle the Tx FEC chunks to avoid error bursts
When sending an FEC-encoded API message, concatenate the chunks from multiple
FEC objects in random order so that the chunks from the same objects are not
consecutive. This strategy avoids error bursts that could affect a single object
and render the message undecodable. By spreading the errors over the multiple
objects, it's more likely that the errors will damage a tolerable number
of (overhead) chunks on each FEC object.