Fix loopback of multicast packets sent on demo-rx
Demo-rx uses the UdpSock objects in Tx mode. The latter prepares the socket for
transmission of multicast-addressed packets. However, UdpSock was previously
using the default configuration for option IP_MULTICAST_LOOP, which is to loop
Tx packets back to the sender. In most cases, the loopback packets are harmless,
but there are specific test scenarios where it's better to disable the
loopback. For example, if running demo-rx to send packets and concurrently
running the listener app on the same interface, and if the goal is to isolate
the two packet streams. Hence, this patch disables the loopback.