Add option to receive the btc source code via API
- Add option --btc-src on the API demo-rx application to allow fetching and
  relaying of packets containing the Bitcoin source code files sent over the
  satellite API.

- Correspondingly, add option --btc-src on the API listener application to
  receive the source code messages.

When the --btc-src option is defined via command-line, configure the
applications to send and listen to API channel 5, to use plaintext mode, and to
connect to the appropriate API server endpoint.

Similar to the --gossip option, use the new --btc-src option to override some
parameters on the demo-rx and listener apps.