Check monitor options on the main sdr app thread
The Monitor object was previously created on a separate thread. Since some
options are parsed when constructing the monitor object, the parsing could fail
inside a thread. As a result, the monitoring thread would break, but the
parent/main thread would continue normally.

To solve the problem, create the Monitor object in the parent thread and pass it
as argument to the monitoring thread. If the construction fails, the parent
thread will fail too.