Access the local config on the Reporter class
Refactor the options passed to the Reporter constructor. Define a common
function to extract the relevant options from the parser object. Use it on all
receivers (standalone, usb, and sdr) for better consistency on the
implementation.

Make sure to include the configuration name and directory within the set of
options so that the reporter class can access the configuration
file. Furthermore, now that the configuration file is accessible, take the
satellite info directly from there instead of a user-specified argument.