Separate API GPG keyring config command and method
From the command-line, one can call "blocksat-cli api config", which is a
command to configure the local GPG keyring. Internally, this command now calls a
standalone method called "config_keyring". Meanwhile, other modules can call
"config_keyring" directly, instead of calling the "config" method that is tied
to the "blocksat-cli api config" command.

The difference is that one method (config_keyring) takes normal arguments and
the other (config) takes a Namespace object from argparse as argument. Hence,
the former is more suitable to be used by other modules.

In this process, review also the rules used to infer when a keyring is already
properly configured. Check the new "_is_gpg_keyring_set" implementation directly
on "config_keyring" and return if the local keyring is already set.