Add wrapper to get the setup's network interface
The configuration file determines the type of receiver being used on the
setup. As such, it also determines the network interface associated with the
receiver. Hence, the network interface can be obtained/inferred from the
configuration file.

This patch adds a wrapper on the config module to return the network interface
associated with a given configuration. For now, this is used on the API listener
command. However, the expectation is that it could be used in other places in
the future.