Move API message generation/decoding to module
The generation and decoding logic was previously on api.py. It was sufficient,
but not easy to reuse in other modules. By moving the logic to wrapper functions
on msg.py, we can easily reuse the logic in the future.

This patch moves the generation logic (previously on api.send()) and the
decoding logic (previously on api.listen()) to the msg.py module. It also adds a
unit test case covering several combinations of parameters.