Move API listener to standalone module and test it
A recent commit moved the listener implementation to a standalone function in an
effort to make it reusable and more easily testable. This patch goes a step
further and moves the loop to a class, implemented on a standalone module.

By having the listener loop on a class, we can keep some state on it. The
primary state of interest is the "enabled" state, which was previously
missing. We need this state so that we can start and stop the (otherwise
infinite) listening loop.

The new class implements a stop() method to break the loop and other options
that are convenient for testing. One of them is the "recv_once" option, which
configures the loop to receive a single message and exit. The other option is
the "recv_queue", through which the caller can provide a queue to store messages
received by the loop. These two options are used on a new set of unit test cases
created to test the loop implementation.

Aside from the above new options, the listener loop implementation has been
essentially preserved.