Support Rx reporting to the new monitoring API
Use option --report (available for all types of receivers) to report
receiver metrics to the new Satellite Monitoring API hosted by
Blockstream. If a specific --report-dest is defined, report to this
destination instead.

When reporting to Blockstream's Monitoring API, handle the initial
registration procedure first. Explain to the user how the two-factor
authentication over satellite works as well as the data collected on
reports. In the end, once the registration completes, save the
registered UUID locally on the JSON config file.

When reporting metrics to the Monitoring API, add the necessary fields
for authentication. In contrast, if reporting to a general-purpose
--report-dest, keep the reports with the previous format (without any
authentication info).