Include TLS key-cert on missing Sat API endpoints
The previous implementation was not adding the TLS client cert on the calls to
`/subscribe/` and `/message/:seq_num`. The rationale was that these endpoints
were always public. More recently, however, a new API server instance has been
deployed to assist with the two-factor authentication process used by the new
Monitoring API. This instance requires client cert for the referred
endpoints (`subscribe/` and `message/:seq_num`). Hence, this patch hooks the
client cert and key to them.