Cache USB adapter number on local config file
When the USB receiver is launched, the USB app selects the appropriate DVB-S2
adapter based on user input. Depending on the choice, a different dvbnet
interface number may result. Meanwhile, in the previous implementation, other
modules that need the interface name (e.g., the API module) didn't check the
adapter number and simply guessed the interface named would be dvb0_0 and
dvb0_1. As a result, the guess could be wrong.

One way to solve the problem would be to call the functionality from usb.py to
search for DVB adapters and let the user select the adapter. However, this
solution requires extra input from the user. For instance, when launching
"blocksat-cli api listen", the user would need to select the DVB-S2 UDB adapter,
which is not ideal.

Instead of doing so, this patch simply caches the adapter number on the local
config file. Every time the user launches the USB receiver, the adapter number
is updated on the config file. This works even if the adapter number changes
between boots. Then, when an app like the API app wants the dvbnet interface
name, the adapter number can be read from the local config file directly.