Review gpg keyring verification sequence
In terms of functionality, the new implementation is identical to the
previous. However, the difference is that it avoids unnecessary warnings by
returning early. As soon as it is known that the keyring does not contain two
public keys, there is no need to search for the blocksat pubkey (which would
return warning if not found).