Fix unintended dict reuse on bitcoin config class
With a default value of {} for the constructor's cfg parameter, it turns out
that all objects constructed with the default cfg value will point to the same
dictionary and reuse it. Hence, the dictionary will accumulate the contents from
multiple Cfg objects, which is undesired. Fix this by using a None default value
instead.