Add cache module and use it on the monitoring API
When executing the initial monitoring API registration, use the cache
file to avoid displaying the monitoring API explainer twice.

Take the previous update cache handler as reference and split it into
two classes: one for the update cache, the other a more basic base class
for cache handling in general. Use the latter on the monitoring API
registration logic.