Add option to disable the sat-ip frontend monitor
The CLI authenticates with the Sat-IP server in order to fetch the
DVB-S2 frontend status metrics. However, this fails on old versions of
the Selfsat Sat-IP firmware lacking a login mechanism. This patch adds
an option to disable the frontend monitoring completely, in which case
the login mechanism becomes unnecessary. When the frontend monitoring is
disabled, the firmware verification routine is adapted to allow for
running with an old firmware.