TBS drivers: fall back to default kernel-headers
When installing the kernel-headers package via dnf, if the package
version does not match the current kernel version, suggest a kernel
upgrade first. After that, if it still doesn't match, try with the
default kernel-headers package. Also, if the user denies the kernel
upgrade, try with the default kernel-headers package.