fixup! bc182f2 Find serving frontend once
Find the serving frontend once in the beginning and stick to it when
logging the frontend status. If the serving frontend disappears from the
active frontend list, check again and refresh the serving frontend.

Also, when searching for the serving frontend, check the DVB-S2
parameters first. If only one candidate frontend is found, pick it and
don't filter the client IP address. The latter can be problematic when
running behind a NAT, as noted in the code.