Review CHANGELOG versions and bump to 0.4.0
The confusion regarding versions on the changelog is that they were
previously stated according to the commit tags, which are starting with
"2.x". However, only the tags are starting with "2.x", as they were
incremented from satellite v1 (GNU Radio version) to satellite v2
(standard DVB-S2 version). In contrast, the CLI, which was added on
satellite v2 only, started with its own versioning, but matching the
suffix of "2.x". For instance, tag 2.3.2 was added with the release of
blocksat-cli 0.3.2. Since then, we have been keeping tags and CLI
releases matched, aside from the "2.x" prefix difference.

This patch removes the "2.x" prefix from the changelog to clarify the
blocksat-cli versions that the changelog is actually tracking. The Git
tags will be preserved with the "2.x" prefix to keep the monotonic
increasing version. However, both the CLI and the changelog will now
start with "0." at this point (the CLI version was already like that).