ip: Fix source appending to /etc/network/interface
The previous configuration of the /etc/network/interfaces file would
only work if executed by the root user. This patch adopts a similar
solution as the one used to create a file in a directory requiring root
permissions (using the ProcessRunner.create_file function). Now, the CLI
appends the source line to a temporary file first, which doesn't require
root permissions. Then, it runs a move command back to the original
/etc/network/interface path with sudo in front if necessary.