s400: Fix parsing of lock status read via SNMP
The receiver metrics fetched via SNMP come from independent SNMP
observations, not a single atomic request. Hence, the lock status can
change in the middle of the observations. That breaks the previous
parsing implementation.

The lock status can return True while the other metrics come empty due
to the lock status turning False. The previous implementation would try
parsing the metrics as-is (empty), thinking they are not empty, in which
case exceptions would be thrown when type-casting '' (empty string) to
float. To solve the problem, check if any metric is empty before
proceeding. In the positive case, assume Lock is False even if the SNMP
report says Lock=True.